fix: Do not reexport integration setup functions
